Conflicts:
	src/search-tool/search-tool.cc

+++ src/modplug/settings.h
@@ -22,7 +22,7 @@ typedef struct {
     int mSurroundDelay;
 
     bool mPreamp;
-    float mPreampLevel;
+    double mPreampLevel;
 
     bool mOversamp;
     bool mNoiseReduction;
